Link to Slack signup
Originally, we were to link the button in the About Us section to the
Slack signup, but this is a bit too tucked away. Instead, that button
will also link to the Meetup group and the Slack signup link will appear
in the header navigation.

In addition, I feel that "Connect with Ruby enthusiasts" feels more
encouraging of the kind of community we want to foster as opposed to
just "hanging out".