Gracefully redirect URLs supported by the former bostonrb site
Summary:
This PR introduces two redirects to gracefully steer users away from
paths that are not supported by the new website. As we decide how to
handle presentations, we may decide to remove these redirects and render
content. With that in mind, these redirects use the temporary
redirect status code, 307.