Force SSL
Summary:
The Cloudflare setup includes a shared Cloudflare Universal SSL
certificate. This means the site is accessible over `https`. However, the
assets within the site don’t load because of mixed
content warning as the assets are being requested over `http`.

This PR adds an environment variable to configure the app to force SSL.
`force_ssl` will rewrite the links to use `https` and visiting `http://`
will redirect to `https://`.