Migrate to sassc-rails
Summary:
The ruby-sass library (ie: 'sass') is being deprecated and will be
unmaintained as of March 26, 2019. The deprecation warning and
maintainers of sass [recommend using the sassc gem] instead.
In our case, we want to replace sass-rails with sassc-rails.

What is sass-rails or sassc-rails?
sass-rails and sassc-rails are packages that integrate Sass into the
Rails asset pipeline. This package allows us to write stylesheets using
Sass and have them compiled into CSS.

[recommend using the sassc gem]: http://sass.logdown.com/posts/7081811