Rewrite (#76)
With heroku charging, it's going to be simpler to make this a static
site. The only part of this that requires a server is the part that
hits meetup.com's API to get the next scheduled event. We can
instead move that into a separate service that we can deploy
separately and keep this as a static site.